Add option to ignore 'SizeM' for pyramid subblocks
This commit introduces a new command line option `-i, --ignoresizem` in the `CCmdLineOptions` class, allowing users to specify whether to ignore the 'SizeM' field for pyramid subblocks. A new boolean member variable `ignore_sizem_for_pyramid_subblocks_` is added and initialized to `false`. The parsing logic for this option is implemented, and the `Run` method in `CRunChecks` is updated to utilize this new setting when opening CZI files.